refactor: seperated components
