+++ src/App.css
@@ -36,7 +36,7 @@ h1 {
 
 .App header {
   background-color: #FAFAFA;
-  height: 10vh;
+  height: 9vh;
   min-height: 50px;
   color: black;
   position: fixed;
@@ -46,11 +46,12 @@ h1 {
   align-items: center;
   justify-content: space-between;
   z-index: 99;
-  padding: 10px;
+  padding: 5px;
   box-sizing: border-box;
   font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
   font-size: small;
-  box-shadow: 0 0 20px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+  border-radius: 5px;
+  box-shadow: rgba(0, 0, 0, 0.1) 0px 20px 25px -5px, rgba(0, 0, 0, 0.04) 0px 10px 10px -5px;
 }
 
 .App section {
@@ -58,10 +59,6 @@ h1 {
   flex-direction: column;
   justify-content: center;
   min-height: 90vh;
-  /* background-image: url(./img/undraw_Messaging_fun_re_vic9.svg  );
-  background-repeat: no-repeat;
-  background-size: 70% 70%;
-  background-position: center; */
 }
 
 main {
@@ -83,31 +80,36 @@ main::-webkit-scrollbar-thumb {
 }
 
 form {
-  height: 11vh;
+  height: 9vh;
   position: fixed;
   bottom: 0;
-  width: 100%;
+  width: calc(100% - 0.5rem);
+  margin-left: 0.3rem;
+  margin-bottom: 0.1rem;
+  border-radius: 50px;
+  overflow: hidden;
   display: flex;
   justify-content: space-between;
-  box-shadow: 0px 0px 20px 0px rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+  box-shadow: rgba(0, 0, 0, 0.25) 0px 54px 55px, rgba(0, 0, 0, 0.12) 0px -12px 30px, rgba(0, 0, 0, 0.12) 0px 4px 6px, rgba(0, 0, 0, 0.17) 0px 12px 13px, rgba(0, 0, 0, 0.09) 0px -3px 5px;
 }
 
 form button {
   width: 10%;
+  border-radius: 0% !important;
 }
 
 
 input {
   line-height: 1.5;
   width: 90%;
-  font-size: 1.5rem;
+  font-size: 1rem;
   background: #f3f4f5;
   color: black;
   outline: none;
   border: none;
   padding: 0 10px;
   font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-
+  border: 0.1px solid rgba(128, 128, 128, 0.096);
 }
 
 button {
